Struct rly_common::builders::BuilderBase

pub struct BuilderBase { /* fields omitted */ }

This is the base struct for the ParserBuilder and LexerBuilder types.

Implementations

impl BuilderBase

pub fn new() -> Self

Creates a new instance of BuilderBase with name and location set to None.

pub fn name(&self) -> Option<&OsStr>

Returns the name used by the builder.

pub fn location(&self) -> Option<&Path>

Returns the location used by the builder.

pub fn with_name<S: Into<OsString>, T: Display, U: Display>(
&mut self,
name: S,
module_type: T,
file_extension: U
) -> Result<&mut Self, Error>

Attempts to set the value of self.name to name.

If name has no path components and no file extension, then the method changes self.name accordingly and returns Ok(self). Otherwise returns an Error with a message depending in module_type and file_extension.

pub fn at_location<S: Into<PathBuf>, T: Display, U: Display>(
&mut self,
location: S,
module_type: T,
file_extension: U
) -> Result<&mut Self, Error>

Attempts to set the value of self.location to location.

If location is a Path to an existing file with extension file_extension, then the method changes self.location accordingly and returns Ok(self). Otherwise returns an Error with a message depending in module_type and file_extension.

pub fn set_name<T: AsRef<OsStr>>(&mut self, end: T)

Sets self.name to a default value.

The default value is self.location if set, and the name of the package being built package plus “_<end>” if not.

pub fn set_path<S: AsRef<OsStr> + ?Sized>(&mut self, ext: &S)

Sets self.name to a default value.

The defulat value is [self.name] plus the extension ext.

Panics

Panics if self.name is None.

Example

use rly_common::builders::BuilderBase;
let mut builder = BuilderBase::new();
builder.set_path("example")             // panics because 'builder.name' is 'None'

pub fn set_rerun_for_location(&self) -> Option<()>

Tells cargo to watch self.location for changes and rerun the build script if it detects any.

Returns Some if self.location is set and the method therefore is successful. If self.location is not set, it returns None and cargo is not set to watch anything for changes.