Module rly_common::errors

A module for generic errors found in text.

This module contains two structs: ErrorData and LineOffset. ErrorData contains information usefull for displaying error messages about the location of an error within a text source, including the line number the error occured on, the collumn it occured on, the index of the error within the source, and the line on which it occured.

An ErrorData can be generated from the source text and an index or position within that text using ErrorData::find, like so:

let text = "\
abc def
ghi jkl
mnop";

let err_data = ErrorData::find(text, 12).unwrap();

assert_eq!(err_data.line(), "ghi jkl");
assert_eq!(err_data.lineno(), 1);
assert_eq!(err_data.colno(), 4);
assert_eq!(err_data.pos(), 12);

LineOffset is essentially just a wrapper for ErrorData that implements Display by printing the line on which the error occured along with a caret ^ inserted underneath the beginning of the error, like so:

let text = "\
abc def
ghi jkl
mnop";

let err_data = ErrorData::find(text, 12).unwrap();

assert_eq!(&err_data.line_offset().to_string(), "\
ghi jkl
    ^  ");

Structs

ErrorData

A type containing information about the location of an error in text.

LineOffset

A struct that Displays its line with a caret ^ underneath the collumn colno.